refactor: separate lambda handler and core modules
